Indicator: Mortalidade por causas externas

Indicator: Mortalidade por causas externas

Usage

indi_0001(
  agg,
  agg_time = "year",
  ano,
  multi = 1e+05,
  decimals = 2,
  complete_with_zeros = TRUE,
  keep_raw_values = FALSE,
  save_args = FALSE,
  pcdas_token = NULL
)

Arguments

agg

character. Spatial aggregation level. uf_res for UF of residence. uf_ocor for UF of occurrence. regsaude_res for regiao de saude of residence. regsaude_ocor for regiao de saúde of occurence. mun_res for municipality of residence. mun_ocor for municipality of ocurrence.

agg_time

character. Time aggregation level. year for yearly data. month for monthly data. week for weekly data. Defaults to year.

ano

numeric. Year of death.

multi

integer. Multiplicator for indicator.

decimals

integer. Number of decimals for indicator.

complete_with_zeros

logical. Complete indicator result with zeros considering combinations of spatial and temporal aggregation without results.

keep_raw_values

logical. Keep numerator and denominator values on results.

save_args

logical. Save agg and agg_time arguments on results table.

pcdas_token

character. PCDaS API token. If not provided, the function will look for it on renvirom.